requirements.txt
aiofiles
aiohttp
aiopytesseract
aiosignal
aio-mc-rcon
apscheduler
async-timeout
attrs
certifi
charset-normalizer
ffmpeg-python
frozenlist
idna
multidict
pillow
py-cord
requests
typing_extensions
urllib3
wget
yarl
youtube_transcript_api 
pytubefix